+++ info/build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.4'
 }
 
-group = 'com.example'
+group = 'com.popup.info'
 version = '0.0.1-SNAPSHOT'
 
 java {
@@ -23,7 +23,7 @@ dependencies {
 	implementation ('org.springframework.boot:spring-boot-starter-data-mongodb')
 	implementation('org.projectlombok:lombok')
 	annotationProcessor ('org.projectlombok:lombok')
-
+	implementation files('/Users/suhyun/Crawler/info/build/libs/domain-1.0-PopUpStoreInfo.jar')
 }
 
 tasks.named('bootBuildImage') {
